==> Scripts/01 KU Importing Libraries and Python Data Types.ipynb <==
# Python Data Types

# Import libraries 
import pandas as pd
import numpy as np
import os

99 + 1

5 * 20

103 - 3 

x = 1.1

y = 2.2

z = 4.4

y / x

z / x

m = 'Fu'

n = 'tur'

o = 'ama'

m + n + o

l = 'All new, '

p = '! '

q = 'Tonight at seven! '

l + m + n + o + p + q
