==> Scripts/02 KU Data Import and Descriptive Analysis.ipynb <==
# Descriptive Analysis

Import Libraries

# Import libraries 
import pandas as pd
import numpy as np
import os

Import Dataset

# Import csv orders
df = pd.read_csv(r'C:\Users\kacee\OneDrive\Desktop\Instacart Basket Analysis\02 Data\Original Data\orders.csv', index_col = False)

Create path shortcut

# path shortcut
path = r'C:\Users\kacee\OneDrive\Desktop\Instacart Basket Analysis'

path 

df = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), index_col = False)

Create vars_list

vars_list = ['order_id', 'user_id', 'order_number', 'order_dow', 'order_hour_of_day', 'days_since_prior_order']

vars_list

Create subset with vars_list

df = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), usecols = vars_list)

df.head()

Import Dataset

# Import csv products
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = False)

df_prods.head(20)

df_prods.tail(35)

df_prods.columns

df_prods.shape

df_prods.describe()

Step 11, Question 1: Max valuse for aisle_id is 134.000000

df_prods.dtypes

Step 11, Question 2: department_id is an integer
