==> Scripts/04 KU Data Wrangling- Orders.ipynb <==
# Data Wrangling

Import Libraries

# Import libraries 
import pandas as pd
import numpy as np
import os

Import Datasets

# Create path shortcut
path = r'C:\Users\kacee\OneDrive\Desktop\Instacart Basket Analysis'

# Import csv orders
df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), index_col = False)

# Import csv products
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = False)

Clean Orders Dataset- Drop and rename columns, change data types

# Dropping column 'eval_set' from df_ords
df_ords = df_ords.drop(columns = ['eval_set'])

# Renaming column order_dow to orders_day_of_week
df_ords.rename(columns = {'order_dow' : 'orders_day_of_week'}, inplace = True)

df_ords.head()

# Change order_id in df_ords to string
df_ords['order_id'] = df_ords['order_id'].astype('str')

df_ords['order_id'].dtype

df_prods['department_id']==19

df_prods[df_prods['department_id']==19]

# labeling department_id 19 as df_snacks
df_snacks =  df_prods[df_prods['department_id']==19]

df_snacks.head()

# Exercise 4.4

Continue to clean Orders Dataset

# Change user_id in df_ords to string
df_ords['user_id'] = df_ords['user_id'].astype('str')

df_ords['user_id'].dtype

# Renaming column days_since_prior_order to days_since_last_order
df_ords.rename(columns = {'days_since_prior_order' : 'days_since_last_order'}, inplace = True)

df_ords.head()

df_ords['order_hour_of_day'].value_counts()

Import Department Dataset

# Importing department data
df_dep = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'departments.csv'), index_col = False)

Transpose Department Dataset from wide to long format

# Transposing df_dep
df_dep.T

# Renaming df with transposed data to df_dep_t
df_dep_t = df_dep.T

# Reset the index 
df_dep_t.reset_index()

# Make first row of df_dept_t the new header
new_header = df_dep_t.iloc[0]

# New DF taking out old header at row 0
df_dep_t_new = df_dep_t[1:]

df_dep_t_new

#set header to column names
df_dep_t_new.columns = new_header

df_dep_t_new

# department_id data dict
data_dict = df_dep_t_new.to_dict('index')

data_dict

Step 5 answer: department 4 is produce

df_prods[df_prods['department_id']==14]

# Subset breakfast items
df_breakfast_items = df_prods[df_prods['department_id']==14]

df_breakfast_items

# Subset dinner party items
df_dinner_party_items = df_prods.loc[df_prods['department_id'].isin([17,18,19])]

df_dinner_party_items

Step 8 answer: df_dinner_party_items has 10430 rows

#Information df on user_id 1
df_userid_1 = df_ords.loc[df_ords['user_id']=='1']

df_userid_1

df_userid_1.describe()

# Exporting df_ords
df_ords.to_csv(os.path.join(path, '02 Data','Prepared Data', 'orders_wrangled.csv'))

df_ords.shape
