==> Scripts/05 KU Data Consistency Check- Orders and Products.ipynb <==
# Checking Data For Consistency

Import Libraries

# Import libraries 
import pandas as pd
import numpy as np
import os

Import Datasets

# Create path shortcut
path = r'C:\Users\kacee\OneDrive\Desktop\Instacart Basket Analysis'

# Import csv products
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = False)

# Import csv orders
df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'orders_wrangled.csv'), index_col = False)

Create a Data frame to test mixed data

# Create a data frame to test mixed data
df_test = pd.DataFrame()

# Create a mixed type column
df_test['mix'] = ['a', 'b', 1, True]

df_test.head()

# Checking for mix data types in test
for col in df_test.columns.tolist():
  weird = (df_test[[col]].applymap(type) != df_test[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_test[weird]) > 0:
    print (col)

# Make mix column string
df_test['mix'] = df_test['mix'].astype('str')

Consistency check for Products

# Checking for Missing Values in df_prods
df_prods.isnull().sum()

#Create subset with missing values from product_name in df_prods
df_nan = df_prods[df_prods['product_name'].isnull() == True]

df_nan

df_prods.describe()

df_prods.shape

#Create new df_prods by removing missing values: df__prods_clean
df_prods_clean = df_prods[df_prods['product_name'].isnull() == False]

df_prods_clean.shape

#Create new df for duplicates from df_prods_clean
df_dups = df_prods_clean[df_prods_clean.duplicated()]

df_dups

#Create new df_prods_clean with no duplicatesL df_prods_clean_no_dups
df_prods_clean_no_dups = df_prods_clean.drop_duplicates()

df_prods_clean_no_dups.shape

df_prods_clean_no_dups

# Exercise 4.5

Consistency check for Orders

df_ords.describe()

Question 2 Answers

Order number: min 1, max 100, 50% 11
Orders day of week: min 0, max 6, 50% 3
Oder hour of day: min 0, max 23, 50% 13
Days since last order: min 0, max 30, 50% 7

Order number max seems really high.

# Checking for mix data types in df_ords
for col in df_ords.columns.tolist():
  weird = (df_ords[[col]].applymap(type) != df_ords[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_ords[weird]) > 0:
    print (col)

df_ords.dtypes

df_ords.head()

df_ords = df_ords.drop(columns = ['Unnamed: 0'])

df_ords

#Check for missing values in df_ords
df_ords.isnull().sum()

Question 5 Answers

days_since_last_order has 206209 missing values

#Create subset with missing values from days_since_last_order in df_ords
df_nan_ords = df_ords[df_ords['days_since_last_order'].isnull() == True]

df_nan_ords

df_nan_ords.describe()

Question 6 Answers

A subset was created named df_nan_ords to show all missing values from days_since_last_order. 
Every single entry shows that this is order number 1 (their first order), so it makes since they have no days since their last order. 
NaN will remain unchanged.

df_ords.shape

#Create new df for duplicates from df_ords
df_dups_ords = df_ords[df_ords.duplicated()]

df_dups_ords

Question 7 Answers

There are no duplicates in df_ords.

Export Wrangled and Prepped Datasets

# Rename and Export df_ords
df_ords.to_csv(os.path.join(path, '02 Data','Prepared Data', 'orders_prepped.csv'))

# Rename and Export df_prods
df_prods_clean_no_dups.to_csv(os.path.join(path, '02 Data','Prepared Data', 'products_prepped.csv'))
