==> Scripts/06 KU Merge Data- Orders and orders_products_prior.ipynb <==
# Merging Dataframes

Import Libraries

# Import libraries 
import pandas as pd
import numpy as np
import os

Import Datasets

# Create path shortcut
path = r'C:\Users\kacee\OneDrive\Desktop\Instacart Basket Analysis'

# Import csv orders_products_prior
df_ords_prior = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders_products_prior.csv'), index_col = False)

# Import csv orders
df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'orders_prepped.csv'), index_col = False)

# Check output ords_prior
df_ords_prior.head()

Prepare Datasets for Merge

# Checking for Missing Values in df_ords_prior
df_ords_prior.isnull().sum()

#Check output ords
df_ords.head()

df_ords_prior.shape

df_ords.shape

df_ords = df_ords.drop(columns = ['Unnamed: 0'])

df_ords.head()

df_ords.shape

Merge Datasets

# Merging df_ords and df_ords_prior inner
df_merged_large = df_ords.merge(df_ords_prior, on = 'order_id', indicator = True)

# Check output
df_merged_large.head()

# Check for value match
df_merged_large['_merge'].value_counts()

# Check Shape
df_merged_large.shape

Export Dataset

# Export df in pickle format
df_merged_large.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_combined.pkl'))
