==> Scripts/07 KU Merge Data- Orders and Products- Large Dataset.ipynb <==
# Exercise 4.6

Import libraries and create path shortcut

# Import libraries 
import pandas as pd
import numpy as np
import os

# Create path shortcut
path = r'C:\Users\kacee\OneDrive\Desktop\Instacart Basket Analysis'

Import orders_products_combined.pkl

# Import pkl orders_products_combined
df_ords_prods_comb = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_combined.pkl'))

Check shape of imported df

# Check Shape ords_prods_comb
df_ords_prods_comb.shape

 Check data sets and make alterations before merging dataframes

# Import csv products
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'products_prepped.csv'), index_col = False)

# Check shape prods
df_prods.shape

# Check output ords_prods_comb
df_ords_prods_comb.head()

# Check output prods
df_prods.head()

# Drop unnamed column prods
df_prods_new = df_prods.drop(columns = ['Unnamed: 0'])

# Check output prods_new
df_prods_new.head()

# Drop _merge column ords_prods_comb
df_ords_new = df_ords_prods_comb.drop(columns = ['_merge'])

# Check output ords_new
df_ords_new.head()

Merging the two data frames orders_products_combined and products_prepped

# Merging the ords_prods_combined df with the df_prods on = product_id column
df_ords_prods_merged = df_ords_new.merge(df_prods_new, on = 'product_id', indicator = True)

Confirm the results with a merge flag

# Check df_ords_prods_merged output
df_ords_prods_merged.head()

# Check for value match
df_ords_prods_merged['_merge'].value_counts()

# Check output
df_ords_prods_merged

Export Large Dataset

# Export new merged df in pickle format
df_ords_prods_merged.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_merged.pkl'))
