==> Scripts/08 KU Deriving New Variables for Large Dataset.ipynb <==
# Deriving New Variables 

Import Libraries

# Import libraries 
import pandas as pd
import numpy as np
import os

Import Dataset

# Create path shortcut
path = r'C:\Users\kacee\OneDrive\Desktop\Instacart Basket Analysis'

# Import pkl orders_products_merged
df_merged = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_merged.pkl'))

Create subset of first million rows

#Create subset
df = df_merged[:1000000]

# Check shape of subset 
df.shape

# Check output of subset
df.head()

Create a User-Defined function to determine price range

# Create price_label function

def price_label(row):

  if row['prices'] <= 5:
    return 'Low-range product'
  elif (row['prices'] > 5) and (row['prices'] <= 15):
    return 'Mid-range product'
  elif row['prices'] > 15:
    return 'High range'
  else: return 'Not enough data'

# Apply price_label function to subset
df['price_range'] = df.apply(price_label, axis=1)

# Check value counts for price_range
df['price_range'].value_counts(dropna = False)

# Check for max value in subset
df['prices'].max()

Creating price_range_loc with loc function in subset

df.loc[df['prices'] > 15, 'price_range_loc'] = 'High-range product'

df.loc[(df['prices'] <= 15) & (df['prices'] > 5), 'price_range_loc'] = 'Mid-range product'

df.loc[df['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

# Check value counts for price_range_loc
df['price_range_loc'].value_counts(dropna = False)

Creating price_range_loc with loc function in df_merged

df_merged.loc[df_merged['prices'] > 15, 'price_range_loc'] = 'High-range product'

df_merged.loc[(df_merged['prices'] <= 15) & (df_merged['prices'] > 5), 'price_range_loc'] = 'Mid-range product'

df_merged.loc[df_merged['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

# Check value counts for price_range_loc
df_merged['price_range_loc'].value_counts(dropna = False)

Creating loop for orders_day_of_week

result = []

for value in df_merged["orders_day_of_week"]:
  if value == 0:
    result.append("Busiest day")
  elif value == 4:
    result.append("Least busy")
  else:
    result.append("Regularly busy")

# Check loop result
result

df_merged['busiest_day'] = result

# Check values for loop busiest_day
df_merged['busiest_day'].value_counts(dropna = False)

# Change df name
df_ords_prods_merged = df_merged

# Exercise 4.7

Create busiest_days column

# Create loop for busiest days

result_2 = []

for value in df_ords_prods_merged['orders_day_of_week']:
  if value == 0 or value == 1:
    result_2.append('Busiest days')
  elif value == 4 or value == 3:
    result_2.append('Least busy days')
  else:
    result_2.append('Regularly busy')

# Check loop result
result_2

# Add result_2 to df in new column
df_ords_prods_merged['busiest_days'] = result_2

# Check output
df_ords_prods_merged.head()

Check values for busiest_days column

# Check values for loop busiest_days
df_ords_prods_merged['busiest_days'].value_counts(dropna = False)

Create column for busiest hours of the day

df_ords_prods_merged['order_hour_of_day'].value_counts(dropna = False)

# Creat loop for busiest hours

result_3 = []

for value in df_ords_prods_merged['order_hour_of_day']:
  if value == 10 or value == 11 or value == 14 or value == 15 or value == 13 or value == 12 or value == 16 or value == 9:
    result_3.append('Most orders')
  elif value == 3 or value == 4 or value == 2 or value == 5 or value == 1:
    result_3.append('Fewest orders')
  else:
    result_3.append('Average orders')

# Check loop result
result_3

# Create new column busiest_period_of_day
df_ords_prods_merged['busiest_period_of_day'] = result_3

Print frequency for the new column

# Check values for loop busiest_period_of_day
df_ords_prods_merged['busiest_period_of_day'].value_counts(dropna = False)

Export Dataset

# Export df in pkl format
df_ords_prods_merged.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_merged_updatedcolumns.pkl'))
