==> Scripts/10 KU Grouping Data- Loyalty Flag.ipynb <==
# Grouping Data & Aggregating Variables

Import Libraries

# Import libraries 
import pandas as pd
import numpy as np
import os

Import Datasets

# Create path shortcut
path = r'C:\Users\kacee\OneDrive\Desktop\Instacart Basket Analysis'

# Import pkl orders_products_merged_updatedcolumns
df_ords_prods_merge = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_merged_updatedcolumns.pkl'))

# Check output
df_ords_prods_merge

Cleanup Dataset

# Remove _merge column
df_ords_prods_merge = df_ords_prods_merge.drop(columns = ['_merge'])

Creating subset to explore data

# Create subset
df = df_ords_prods_merge[:1000000]

# Check subset shape
df.shape

# Check subset output
df.head()

# Group subset by product name
df.groupby('product_name')

# Average number of orders per user across each Instacart department using agg()
df.groupby('department_id').agg({'order_number': ['mean']})

# Average number of orders per user across each Instacart department
df.groupby('department_id')['order_number'].mean()

# Mean, Min, Max orders across each Instacart department
df.groupby('department_id').agg({'order_number': ['mean', 'min', 'max']})

Create max_order column and create a loyalty_flag in complete dataset

# New column- Max order per customer
df_ords_prods_merge['max_order'] = df_ords_prods_merge.groupby(['user_id'])['order_number'].transform(np.max)

# Check output with max order column
df_ords_prods_merge.head(15)

# Clear max number of rows that can be displayed
pd.options.display.max_rows = None

# Create a loyalty flag
df_ords_prods_merge.loc[df_ords_prods_merge['max_order'] > 40, 'loyalty_flag'] = 'Loyal customer'
df_ords_prods_merge.loc[(df_ords_prods_merge['max_order'] <= 40) & (df_ords_prods_merge['max_order'] > 10), 'loyalty_flag'] = 'Regular customer'
df_ords_prods_merge.loc[df_ords_prods_merge['max_order'] <= 10, 'loyalty_flag'] = 'New customer'

# Count loyalty_flag values
df_ords_prods_merge['loyalty_flag'].value_counts(dropna = False)

# Check loyalty_flag output
df_ords_prods_merge.head(20)

# Check output of 40 rows 
df_ords_prods_merge[['user_id', 'loyalty_flag', 'order_number']].head(40)

Export Dataset

# Export df in pkl format
df_ords_prods_merge.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_merged_groupby.pkl'))
