==> Scripts/12 KU Data Wrangling- Customers.ipynb <==
# Exercise 1 Part 1

Import Libraries

# Import Libraries
import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy

Import Datasets

# Create path shortcut
path = r'C:\Users\kacee\OneDrive\Desktop\Instacart Basket Analysis'

# Import customer data set
df_cust = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'customers.csv'))

# Import pkl orders_products_merged_aggregated data set
df_ords_prods = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_merged_aggregated.pkl'))

# Check df_ords_prods output
df_ords_prods.head(20)

# Check customer output
df_cust.head(20)

# Check shape df_ords_prods
df_ords_prods.shape

# Check shape df_cust
df_cust.shape

Data Wrangling Customer Data Set

#Create new df for duplicates from df_cust
df_dups = df_cust[df_cust.duplicated()]

# Check duplicates df
df_dups.head()

# Renaming column STATE to State
df_cust.rename(columns = {'STATE' : 'State'}, inplace = True)

# Renaming column n_dependants to number_of_dependants
df_cust.rename(columns = {'n_dependants' : 'number_of_dependants'}, inplace = True)

# Dropping columns 'first_name' from df_ords
df_cust_2 = df_cust.drop(columns = ['First Name'])

# Dropping columns 'last_name' from df_ords
df_cust_3 = df_cust_2.drop(columns = ['Surnam'])

# Check output
df_cust_3.head(20)

# Change user_id to string
df_cust_3['user_id'] = df_cust_3['user_id'].astype('str')

# Check values for State
df_cust_3['State'].value_counts()

# Check values for Gender
df_cust_3['Gender'].value_counts()

# Check values for Age
df_cust_3['Age'].value_counts()

# Check values for date_joined
df_cust_3['date_joined'].value_counts()

# Check values for fam_status
df_cust_3['fam_status'].value_counts()

# Check data description
df_cust_3.describe()

# Check data types
df_cust_3.dtypes

# Checking for mix data types
for col in df_cust_3.columns.tolist():
  weird = (df_cust_3[[col]].applymap(type) != df_cust_3[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_cust_3[weird]) > 0:
    print (col)

# Checking for Missing Values
df_cust_3.isnull().sum()

# Check Shape of df_cust_3
df_cust_3.shape

Data Wrangling orders_products_merged_aggregated 

# Change user_id to string
df_ords_prods['user_id'] = df_ords_prods['user_id'].astype('str')

Merge Data Sets

# Combine customers data set and orders_products_merged_aggregated on user_id
df_merged = df_cust_3.merge(df_ords_prods, on = "user_id", indicator = True)

# Check out put of merged data
df_merged.head(20)

# Check Shape
df_merged.shape

Export Complete Dataset

# Export data to pkl

df_merged.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'order_products_customers_merged.pkl'))
